services/pm-agent/pyproject.toml
[project]
name = "pm-agent"
version = "0.1.0"
description = "Woorung-Gaksi Project Manager Agent (Brain)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "langgraph>=0.0.10",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
